+++ core-ng-search-test/src/main/java/org/elasticsearch/plugins/LocalPluginsService.java
@@ -28,7 +28,7 @@ public class LocalPluginsService extends PluginsService {
                     "classpath plugin",
                     "NA",
                     Version.CURRENT,
-                    "17",
+                "21",
                     pluginClass.getName(),
                     null,
                     List.of(),